habits-tracker/requirements.txt
djangorestframework
drf-nested-routers
